cas_dam_admin/requirements.txt
beautifulsoup4==4.7.1
cachetools==3.1.1
certifi==2019.6.16
chardet==3.0.4
Django==2.2.2
django-picklefield==2.0
djangorestframework==3.9.4
google==2.0.2
google-api-python-client==1.7.9
google-auth==1.6.3
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.0
httplib2==0.13.0
idna==2.8
oauthlib==3.0.1
pyasn1==0.4.5
pyasn1-modules==0.2.5
pytz==2019.1
requests==2.22.0
requests-oauthlib==1.2.0
rsa==4.0
six==1.12.0
soupsieve==1.9.1
sqlparse==0.3.0
uritemplate==3.0.0
urllib3==1.25.3
google-api-python-client
python-decouple